Improve Code Quality
Made the code more cleaner and more readable.

Breakdown methods into different level of abstactions for better understanding
