+++ src/main/java/duke/gui/MainWindow.java
@@ -1,8 +1,7 @@
-package duke;
+package duke.gui;
 
-import javafx.application.Application;
+import duke.Duke;
 import javafx.fxml.FXML;
-
 import javafx.scene.control.Button;
 import javafx.scene.control.ScrollPane;
 import javafx.scene.control.TextField;
