+++ src/test/java/duke/DateParserTest.java
@@ -6,6 +6,8 @@ import java.time.LocalDate;
 
 import org.junit.jupiter.api.Test;
 
+import duke.parser.DateParser;
+
 public class DateParserTest {
 
     @Test
